+++ src/main/kotlin/lotto/service/LottoService.kt
@@ -1,0 +1,19 @@
+package lotto.service
+
+import lotto.domain.Lotto
+import lotto.domain.LottoTicket
+import java.math.BigDecimal
+import java.math.RoundingMode
+
+class LottoService(
+    private val lotto: Lotto,
+) {
+    fun lottoIssuance(lottoTicket: LottoTicket): List<LottoTicket> {
+        return (1..lotto.buyCount()).map { lottoTicket }
+    }
+
+    fun calculateProfitRate(totalPrize: Int): BigDecimal {
+        return totalPrize.toBigDecimal()
+            .divide(lotto.money.toBigDecimal(), 2, RoundingMode.DOWN)
+    }
+}
