+++ src/main/kotlin/lotto/view/InputView.kt
@@ -1,0 +1,4 @@
+package lotto.view
+
+class InputView {
+}
