+++ src/main/kotlin/lotto/view/ResultView.kt
@@ -1,0 +1,4 @@
+package lotto.view
+
+class ResultView {
+}
